Easy shape bit of ImmutableArcInst could be invalid after change of foundry
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@5929 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/technology/ArcProto.java
@@ -406,7 +406,7 @@ public class ArcProto implements Comparable<ArcProto>
 	 * The offset amount is the difference between the diffusion width and the overall width.
 	 * @return the width offset of this ArcProto in grid units.
 	 */
-    public long getMaxLayerGridOffset() { return maxLayerGridOffset; }
+    public int getMaxLayerGridOffset() { return maxLayerGridOffset; }
     
     /*
 	private Pref getArcProtoAntennaPref()
